refactor: move EncounterDto from class to interface
Handling the retrieval of encounter data we have found more easy
to handle it as an interface. We have changed previous code to
move towards this approach.
As a conclussion of this, the Serializer should store on LocalStorage
something pretty similar to an EncounterDto, so when we recover it in
byUlid method we only have to parse as a JSON