feat: merged feature/integrate-create-encounter-use-case
We are now building the edit encounter case. First of all, we need to
load the encounter data. Now it will be a very simple version, with only
id (to have an unique identifier) and the name of the encounter. With this,
we can load and show the form (in the next step we will integrate the encounter
load with the form)