feat: more integrations on showing monster detail
Second iteration. Now the edit encounter form shows proper encounter info, not
a dummy one.

Detected one thing to ammend. Somehow the xp value of the monsters is retrieved
as a string, forcing the parseInt function in a property that should be a number.

Another thing to think about is the logic of multiplier outside our Domain.