feat: first iteration on Update Encounter level
We need some way to update encounter level with party level.
We started with a command handler knowing that maybe we should
migrate all this logic to a DomainService, but it is a good approach.

In this first iteration we needed all classes to get PartyTresholds and
extend DomainEncounter with an accesor method to get current Encounter level.